replacing int by size_t to make it work on 64-bit architectures (patch by ojnkpjg)

+++ lib/comhttpsocket.h
@@ -30,11 +30,11 @@ public:
 	/** Disconnects the current socket */
 	int Disconnect();
 	/** Reads in a non blocking fashion (ie, selects and polls) for nTimeout seconds */
-	int NBRead(char* pBuffer, int nLen, int* nBytesWritten, int nTimeout);
+	int NBRead(char* pBuffer, size_t nLen, size_t* nBytesWritten, int nTimeout);
         /** Reads from a socket, into pbuffer, up to a max of nLen byte, and writes how many were actually written to nBytesWritten. */
-        int Read(char* pBuffer, int nLen, int* nBytesWritten);
+        int Read(char* pBuffer, size_t nLen, size_t* nBytesWritten);
 	/** Writes to a socket, from buffer pBuffer, up to nLen bytes, and returns the number of written bytes in pnBytesWritten. */
-	int Write(const char* pBuffer, int nLen, int* pnBytesWritten);
+	int Write(const char* pBuffer, size_t nLen, size_t* pnBytesWritten);
 	/** Sets the proxy address. Use NULL to disable. */
 	int SetProxy(const char* pURL);
 
@@ -57,7 +57,7 @@ private: // Private attributes
 	const char* m_pFile;
 	/** temp buffer for excess data when parsing header */
 	char* m_pTempBuf;
-	int m_nBufLen;
+	size_t m_nBufLen;
 };
 
 #endif
